Control aspects of the grid search racing process

Control aspects of the grid search racing process

control_race(
  verbose = FALSE,
  verbose_elim = TRUE,
  allow_par = TRUE,
  extract = NULL,
  save_pred = FALSE,
  burn_in = 3,
  num_ties = 10,
  alpha = 0.05,
  randomize = TRUE,
  pkgs = NULL,
  save_workflow = FALSE
)

Arguments

verbose
A logical for logging results as they are generated. Despite this argument, warnings and errors are always shown. If using a dark IDE theme, some logging messages might be hard to see. If this is the case, try setting the tidymodels.dark option with options(tidymodels.dark = TRUE) to print lighter colors.

verbose_elim
A logical for whether logging of the elimination of tuning parameter combinations should occur.

allow_par
A logical to allow parallel processing (if a parallel backend is registered).

extract
An optional function with at least one argument (or NULL) that can be used to retain arbitrary objects from the model fit object, recipe, or other elements of the workflow.

save_pred
A logical for whether the out-of-sample predictions should be saved for each model evaluated.

burn_in
An integer for how many resamples should be completed for all grid combinations before parameter filtering begins.

num_ties
An integer for when tie-breaking should occur. If there are two final parameter combinations being evaluated, num_ties specified how many more resampling iterations should be evaluated. After num_ties more iterations, the parameter combination with the current best results is retained.

alpha
The alpha level for a one-sided confidence interval for each parameter combination.

randomize
Should the resamples be evaluated in a random order?

pkgs
An optional character string of R package names that should be loaded (by namespace) during parallel processing.

save_workflow
A logical for whether the workflow should be appended to the output as an attribute.